Store/Update a timestamp.
This commit updates things to store a time-stamp when new entries
are added via the web-interface.  This timestamp is updated when
the bookmark is saved too.

If you're doing addition by hand then you'll need to add a "time"
attribute to your entries.

This closes #17.

TODO: Allow sort by date.